requirements.txt
dashscope
python-dotenv
pandas
Pillow
fastapi
uvicorn
python-multipart
